+++ packages/http-client/src/lib/clients/stamps/StampsClient.types.ts
@@ -1,0 +1,24 @@
+import { AxiosInstance } from 'axios'
+import { Newable, Parser } from '../../../types'
+
+export type Avro = {
+  candidate: Record<string, string | number>
+}
+
+export type ClientConfig = {
+  baseUrl: string
+}
+
+export type GetAvroParams = {
+  candid: string
+  survey_id: string
+}
+
+export interface IStampsClient {
+  initClient(axiosInstance?: AxiosInstance): void
+  getAvroJson<T>(
+    params: GetAvroParams,
+    parser?: Parser<Avro, T>,
+    customModel?: Newable<T>
+  ): Promise<T>
+}